factor out maybe_store_exception
I also changed the range used in ExceptionEntry from exc.range() to
stmt.range(), like the currently_suspended_exceptions case below.

this range is not used for the diagnostic but for deduplicating/sorting
exceptions referring to the same type. I think either exc.range() or
stmt.range() will accomplish the same goal, but it seemed slightly preferable to
be consistent in that case.

there's still a bit of duplication here with the same Tuple case in both
branches, but that seems okay.