Avoid extra parentheses for long `match` patterns with `as` captures
Summary
--

This PR fixes #17796 by taking the approach mentioned in
https://github.com/astral-sh/ruff/issues/17796#issuecomment-2847943862 of simply
recursing into the `MatchAs` patterns when checking if we need parentheses. This
allows us to reuse the parentheses in the inner pattern before also breaking the
`MatchAs` pattern itself.

I haven't really resolved the question of whether or not it's okay always to
recurse, but I'm hoping the ecosystem check on this PR might shed some light on
that.

Test Plan
--

New tests based on the issue and then reviewing the ecosystem check here