[ty] Don't provide completions when in `from <module> i` statement
This works principally by copying `from_import_tokens` and tweaking
it a bit to look for `from module <CURSOR>` token sequences. We actually
wind up needing to be a little more careful here to parse `<module>`
correctly to avoid something like `from os.i<CURSOR>` from being treated
as if the `i` is the beginning of the `import` keyword.

Fixes astral-sh/ty#1494