Remove duplicate preview tests for `FURB101` and `FURB103`
Summary
--

These rules are themselves in preview, so we don't need the additional preview
checks on the fixes or the separate preview tests. This has confused me in a
couple of reviews of changes to the fixes.

Test Plan
--

Existing tests, I checked that the diff between the preview and non-empty
snapshots was empty after removing the preview checks and before deleting the
preview tests