Fix syntax error false positive on alternative `match` patterns
Summary
--

Fixes #21360 by using the union of names instead of overwriting them, as Micha
suggested originally on #21104.

This avoids overwriting the `n` name in the `Subscript` by the empty set of
names visited in the nested OR pattern before visiting the other arm of the
outer OR pattern.

Test Plan
--

A new inline test case taken from the issue